requirements.txt
langchain~=0.3.1
requests~=2.32.3
pandas~=2.2.2
beautifulsoup4~=4.12.3
langchain-core~=0.3.8
python-dotenv~=1.0.1
ipython~=8.27.0
langchain-openai~=0.2.1
PyYAML~=6.0.1
streamlit~=1.32.2